extractor: logic improvements.
Only read from one queue (and only issue one "primary" query) at a
time.

Although it would be nice to read from all queues in parallel, and
process all messages in strict timestamp order, this is impractical
because:

- once a queue gets up to the present time, waiting for the next
  message would stall the entire process (we can't handle new
  waveforms until somebody is assigned a new date of birth)... and
  getting around that would require some really nasty tracking of
  which queries are currently in flight and in what order

- for rate limiting, we really want to put the entire program to
  sleep, and not have *any* queries in flight (consuming memory on
  server side) at that time

- most of the time, we want to issue queries for certain queues more
  often than others (in particular, patient metadata queries are
  likely to be more expensive, and we also don't care about polling
  them nearly as often)

- we don't want output handlers to be relying on a strict ordering of
  messages anyway

It might still be worth considering running multiple queries in
parallel, if it can improve performance.  They wouldn't be kept in
lockstep, but rather, we'd only issue queries for multiple queues if
they were "close enough" to each other in time.