archive: add a unified per-record '_properties' file.
This replaces the '_time' file.  It's a JSON file which can be used by
output handlers to store simple bits of per-record persistent state.
API is more than a little bit nasty, but this should be more robust
than implementing the same thing four or five times... and maybe
better performing too.