MessageParser: formalize errors/warnings.
If a mandatory column cannot be parsed, raise a DBSyntaxError.  (The
caller may choose to treat this as fatal or to ignore the message.)

If an optional column is non-null but cannot be parsed, show a
DBSyntaxWarning and set the appropriate field to None.

In both cases, the error/warning message should include the complete
row and the complete original query for reference.