Dispatcher: nack_message: make replaying optional.
In most cases, we don't want to replay old messages until a new
message is acked.  (For most handlers, when a message is nacked, it's
because we're waiting to receive information from another queue
entirely.)  Waveforms are unusual in that we delay processing each
message until we've seen a later message from the same channel.