DWCDB: permit (un)pickling, and sharing between processes.
This is nasty too.  In particular, if a DWCDB object is inherited when
a process forks, we can't use the same connection from both processes.

On the other hand, we do want to keep the same connection open for all
synchronous requests to the same server (identified by servername)
from the same process.

getpid is probably the wrong thing here but I don't know what the
right thing is.