ChildConnector: more efficient tracking of messages in flight.
Instead of keeping a list of every request we submit to the child
process, just keep a counter of how many requests have been submitted.

If we assume that depickling errors can only occur because of
'message' or 'channel' values (and we assume that the child process
never erroneously acknowledges a message it hasn't seen), we don't
need the list - everything we need to know is in the 'messages'
dictionary.