ExtractorQueue: more efficient checking if a message is already known.
If a message hasn't been seen previously, it won't have an entry in
message_info.  Thus, we can avoid keeping a set of seen messages
per-timestamp.  We still need a list of acked messages - but it can
just be a list, not a set.

Also, MessageInfo doesn't need the channel.