ChildConnector: better cleaning-up of child processes.
This is a horrible mess.  At any rate, we do actually want to wait for
all the children (rather than letting "daemon" mode kill them
haphazardly, which breaks profiling among other things.)  To ensure
that's possible, we want to ensure that child processes don't keep
their siblings' pipes open.  To be extra defensive, we also want to
(try to) send an EXIT request in both the exception and non-exception
cases.

We also don't want to send an unnecessary SYNC_RESPONSE if there
weren't any requests pending; meanwhile, if there were requests
pending and they yielded an exception, we don't want to give up on
waiting for all the *other* children.