*Handler: rework time map handling.
Alerts, Enums, and Numerics are all written to temporary log files
with the PHI sequence number and timestamp.  (Once the record is
finalized, we'll go back and calculate the real time of each event.
This is not implemented yet.)

This means we can drastically reduce the inter-queue dependencies at
extraction time.