TimeMap: try to guess the offset for events after the end of the map.
(For example, enums that occur after the last waveform timestamp.)

This is not the best we could possibly do... at least for enums (and
possibly for alerts and some numerics), the SequenceNumber should be
"predictable enough" that we ought to be able to construct an accurate
time map even without waveforms.  Especially if we can assume that
clock adjustments are large.