Archive: get_record: remove sync option.
The use of the "asynchronous" option by the WaveSampleHandler, in
particular, would cause waves to be lost in "unmapped" records.

Parallel dispatching requires mappings to be known before messages are
submitted to the output handlers, so it makes little sense for the
handlers themselves to try to accomodate these "not-yet-mapped"
mapping IDs.  If we ever want to accomodate "real-time" conversion
where the mapping may not be known when the message is received, that
will have to be done in e.g. MappingIDExtractorQueue.