DWCDB: clean up warnings.
If we find unknown wave/enum/numeric/mapping IDs, then warn about it
(but only once for each type.)

In particular, when testing, it's a common mistake to import a
PatientMapping table that only partially covers the data tables; in
addition to generating lots of tiny records for the unassociated
mapping IDs, it's very slow because all of those unassociated mappings
are assigned to the same channel and hence the same worker process.
(This should be fixed too, but anyway...)

So it's useful to have these warnings.

Also, fix the use of an undefined function _log_warning.  Yeah, this
needs a lot of cleanup.