WaveSampleHandler: kludge for ECG baseline.
It appears to be the case that for ECG signals, the baseline is midway
between the ScaleLower and ScaleUpper values, which is also (plus or
minus 1) midway between the CalibrationScaledLower and
CalibrationScaledUpper values.  However, the CalibrationAbsLower and
CalibrationAbsUpper are 0.0 and 1.0 (rather than, say, -0.5 and 0.5).

(Most often the baseline is 8192, and the range is 8092 to 8292 or
something like that.)

This is a horrible kludge, but let's assume that the *gain* is correct
and the *baseline* should be the midpoint between
CalibrationScaledLower and CalibrationScaledUpper, whatever that is.