Periodically stop processing to finalize old records.
When processing a large database, we need to periodically check for
records that have not seen any new data, and must be finalized.  It
doesn't really matter *when* we do this, it just needs to be done at
some time.

Unfortunately, in the normal state of things (using the
ParallelDispatcher), each worker process has its own Archive which is
only up-to-date with respect to records that have been delegated to
that process.  But the individual processes don't know which records
have been delegated to them, and we don't want them to incorrectly
finalize records that they're *not* responsible for; moreover,
somebody needs to finalize the records that are completely idle and
thus haven't been delegated to any worker process.

So, whenever a certain length of time has elapsed in the main loop
(currently 3 hours), we do the following:

- Call Extractor.flush to ensure that each worker process writes its
  Archive state to disk, and to save the current state of the
  extractor.

- Call ParallelDispatcher.shutdown to stop the worker processes and
  wait for them to exit.

- Create a new Extractor, reloading the state from disk.

- Create a new Archive, reloading the state of all archive records
  from disk.

- Call Archive.finalize_before, checking for records that need to be
  finalized (and spawning a process for each.)

- Call Archive.flush, waiting for the finalization processes to finish.

- Resume reading messages from the new Extractor.