WaveSampleHandler: add extra info about ECG signals.
For ECGs, the channel number may be of interest since it tells us
something about which signals were being analyzed for arrhythmias.

It may also be useful to know whether the signal is digitally derived
from another signal (as in the case of limb leads computed from EASI
leads) and whether it has been band-pass filtered.

(Currently, these only apply to ECGs, because only ECG signals are
ever marked as "derived" and only ECG signals indicate the filter
cutoff frequencies.)