WaveSampleHandler: generate unique signal descriptions.
In the (extremely unusual) case where two simultaneous signals have
the same label, we must give them distinct descriptions so that they
can be mapped to distinct signals in the layout header.

It's worth noting that the order of signals is deterministic based on
the WaveAttrs (and two distinct waveforms should never have the same
WaveAttr, and if they did we'd treat them as a single signal anyway.)
However, this doesn't mean that the same signal will have the same
description in consecutive segments, if different signals appear and
disappear at different times.