AlertFinalizer: write alerts to an annotation file.
Alarm annotations include the onset, announcement, and end of the
alarm.  An additional annotation is added if the message or severity
of an alarm changes, or the alarm becomes silenced or un-silenced.

The onset time, if known, should be close to the exact time of the
triggering event.  How precise the other timestamps are is unclear.

Since there may be multiple alarms active at a time, a unique integer
ID is used to identify each alarm and included (enclosed in braces) in
the annotation aux string.