EnumerationValueFinalizer: write beats to an annotation file.
Beat-like annotations (i.e., enumerations with the label "Annot") are
written as WFDB annotations.

Known annotation codes (ValuePhysioIds) are translated to
corresponding WFDB anntyp values, with subtyp used to distinguish
ambiguous cases.  (In the case of "missed beat" and "signals
inoperable", these are translated to NOTE annotations and
distinguished by the aux value.)

The difference between types 148063, 147543, and 147591 is unknown.
These might represent "atrial paced", "ventricular paced", and "dual
paced" in some order; however, all three are labelled as "P".

If the annotation code is unknown, the original code and label are
reported via aux (and we try to guess, based on the first letter of
the label, what the anntyp should be.)