WaveSampleFinalizer: write signal quality to an annotation file.
Paced pulses are recorded as a PACESP annotation at the corresponding
time and channel number.

Intervals of invalid samples are recorded as a pair of NOTE
annotations "(i" and "i)", and intervals of unavailable samples are
recorded as "(u" and "u)".  (In both cases, consecutive or overlapping
intervals are merged into one.)