WaveSampleHandler: new approach to handle out-of-order data.
To ensure that we don't lose data when it arrives out of order, we
need to keep track of what time *intervals* have been written out (not
just assume that everything has been "caught up" to a particular
time.)  Out-of-order waveforms will require additional segment
splitting, but there shouldn't be any data lost provided that all
waveform timestamps "jump" at the same time.

This is also an attempt to clarify the logic and remove some cruft
that was left over from earlier designs for the WaveSampleHandler.
Instead of the word "flushed", use the word "saved" to be less
confusing - we shouldn't say it's "flushed" until archive.flush has
been called.

Finally, this should also ensure that we only ever write to the *end*
of the current segment (once we've written a chunk of data to the
file, that time interval is saved and any further messages that
overlap with that time interval will be ignored.)  This will be
helpful later in calculating checksums.