TimeMap: add limit argument to get_seqnum.
This changes the algorithm for translating timestamps in ambiguous
cases.

For periodics and enums, if the message sequence number is S, the true
time of the event ought to be between S and S+5120.  For aperiodics
and alerts, we don't know that, but our best guess is that the true
time is *close* to the message sequence number, and shouldn't be
greater than S+5120.

So we allow the caller to specify an upper bound, and try to find a
matching sequence number (within a known interval) that is close to
that upper bound without exceeding it.

For periodics and enums, this should always give the right answer for
negative clock adjustments of more than 5.12 seconds.

There might be no known interval matching the given timestamp if there
was a positive clock adjustment that skipped over that timestamp.
(This shouldn't happen if clock adjustments are clean and resolve_gaps
is used - but it might happen if the timestamp is "impossible", for
example if there are two positive adjustments in rapid succession.)
In that case, we somewhat arbitrarily choose to extrapolate from the
following interval (the earliest interval for which the extrapolated
timestamp would be less than the end of that interval.)

There might be no known interval matching the given timestamp if it
occurs before or after all known reference timestamps.  In that case
we extrapolate from the earliest or latest reference timestamp, as
appropriate.

Or there might be a known interval matching the given timestamp, but
it occurs after the specified limit.  In that case our best guess is
that the earliest such interval is the correct one.