AlertFinalizer: better handling of changing messages.
Don't write a status-change annotation at exactly the same time as the
announcement or end annotation.

If there is a status change before the announcement, then use the last
message before the announcement time for the announcement annotation.
Similarly, ignore status changes that occur after the end time.

For example, if we have these four messages:

 23:48:44.190  *** Apnea 0:26
                onset    = 23:48:27.998
                announce = 23:48:49.310

 23:48:49.310  *** Apnea 0:30
                onset    = 23:48:27.998
                announce = 23:48:49.310

 23:48:54.430  *** Apnea 0:30
                onset    = 23:48:27.998
                announce = 23:48:49.310
                end      = 23:49:00.574

 23:49:20.030  *** Apnea 0:30
                onset    = 23:48:27.998
                announce = 23:48:49.310
                end      = 23:49:21.566

Since the label changes at the announcement time, the announcement
annotation should contain the updated message.  The label doesn't
change between the announcement and end time, so there shouldn't be
any status-change annotations.  However, the end time is increased
from the third message to the fourth, so the end annotation should be
placed at the later end time.  Thus we should have the following
annotations:

23:48:27.998   "+{###} *** Apnea 0:26"
23:48:49.310   "<{###} *** Apnea 0:30"
23:49:21.566   ">{###} *** Apnea 0:30"