Avoid use of "is" with literal constants.
Using "is" with a literal constant is probably wrong and Python 3.8
will whine about it.

We do want _boolean to reject values other than the integer 0 and 1
and the boolean False and True - for example, _boolean(1.0) should
raise a TypeError, even though 1.0 == 1 and 1.0 == True.  So use a
bitwise or, which only works for integer and boolean types.