Handle timestamp strings produced by datetime.__str__.
Timestamp strings retrieved from MS SQL, or from BCP files, use
millisecond precision, and have a space between the milliseconds and
the timezone.  This is the format that the T class uses for formatting
timestamps.

However, there's a bug somewhere in the TimeMap class that causes it
to sometimes store timestamps as datetime objects rather than T, which
uses either second or microsecond precision, and omits the space
before the timezone.  In order to parse the resulting _phi_time_map
files, we need to accept those formats as well.  This should probably
be fixed in TimeMap but for now we need to be able to parse the
existing files.