Stricter parsing of event log files.
When parsing an event log file, we need to distinguish between data
records (containing multiple tab-separated columns), timestamps
(consisting of only ASCII digits), and sequence numbers (consisting of
an 'S' followed by ASCII digits).

Previously, to detect a timestamp, we would simply pass the input
string to the builtin int function and rely on it raising a ValueError
if the input wasn't a decimal number.

However, occasionally, an enumeration value can have both an empty
"label" and an empty "value" - resulting in a data record like
"\t123456\t\n", which Python accepts as a valid integer constant.

(This would result in a "failed to finalize" error, because the
ValuePhysioId isn't ever large enough to be misinterpreted as a
timestamp, though in principle it would be possible.)

Conceivably there could also be numeric values that have an empty
"value", empty "units", and a "label" consisting of an ASCII integer
or an 'S' followed by an ASCII integer.

To avoid incorrectly parsing such data records as timestamps or
sequence numbers, check the line against a regular expression rather
than relying on the error behavior of int().