Add bulk-extract and bulk-verify scripts.
bulk-extract is used to extract large volumes of data from SQL Server;
we use this script to automatically extract and archive a 24-hour
chunk of data each day.

bulk-verify is used to sanity-check the resulting data files and
produce a manifest file, which lists the hash and number of rows for
each table.