For aperiodics, use sequence number as observation time.
For "periodic" numerics, assume that the timestamp in the
NumericValueMessage provides a precise observation time (once it is
translated via the time map.)  Note that btbHR is considered
"periodic".

For "aperiodic" numerics, ignore the timestamp and just use the
sequence number as a coarse observation time.

This is possibly still wrong.