Fix formatting of numbers and UUIDs in BCP output.
When writing out a BCP message, each column value should be written in
the same format as was used originally by freebcp, if possible; this
makes it easier to compare the output with the original input files.

In the case of UUIDs, write them in uppercase; this is the way all of
the columns in the Alert/EnumerationValue/NumericValue/WaveSample
tables are written by freebcp.  (Only patient ID columns are written
in lowercase.)

In the case of Decimals, the values '1.000' and '0.000' were
previously written as '1' and '0' because they are "equal" to True and
False.