bcpdstfix: new module.
This script can be used to heuristically repair BCP dump files that
were obtained from a buggy version of DWC.

The problem is that when the system transitions from local summer time
to local winter time (meaning there is a two-hour interval when the
local wall clock reads "01:XX"), DWC sometimes records these
timestamps incorrectly in the SQL database.

Specifically, timestamps that fall within the final hour of local
summer time, which should be stored as "01:XX:YY.ZZZ -04:00", are
instead stored as "01:XX:YY.ZZZ -05:00".

This does not happen *universally* - for some reason, there are
usually a handful of patients for whom the timestamps are stored
correctly.  However, it does happen for the overwhelming majority of
patients in our deployment.

Note that this bug is dependent on the local timezone settings of the
monitoring system, not the local timezone settings of the DWC export
host nor those of the SQL server.

This bug affects both retrospective timestamps (e.g. "TimeStamp") and
prospective timestamps (e.g. "EndTime") in the Alert,
EnumerationValue, NumericValue, and WaveSample tables.  It does not
affect the PatientMapping table.  It doesn't appear to affect the
Patient, PatientDateAttribute, or PatientStringAttribute tables
(though it's hard to be sure.)

(Curiously, this means that the bug affects the columns that are
spelled "TimeStamp" and not the ones that are spelled "Timestamp".)

The present script heuristically fixes these buggy timestamps by
looking at the difference between the event timestamp and the
corresponding PatientMapping timestamp.  If the event timestamp is
within the ambiguous window and is more than 30 minutes after the
PatientMapping timestamp, it is assumed to be incorrect.

Note that this means the script needs to be able to inspect the
PatientMappings for a couple of hours before/after the ambiguous time
window.  For example, to repair the file Alert.20211106_20211107 or
WaveSample.20211107_20211108, the script must be given
PatientMapping.20211106_20211107 as well as
PatientMapping.20211107_20211108 for context.

All of this needs to be done as a preprocessing step because downcast
itself assumes that messages are presented "more or less in
chronological order".