BCPTable: keep open file descriptors for all loaded data files.
Once a BCP "connection" has been opened, it is preferable to keep open
file descriptors to the data files, and reuse those FDs each time a
cursor is needed, so that if the files are deleted/renamed out from
under us, the existing "connection" keeps working.

However, we do *not* simply want to dup the file descriptors - it is
possible and permissible to perform multiple queries on the same data
files in parallel, which means each cursor needs to have its own
independent file object(s) referring to its own independent OFD(s).
On Linux, we can get this behavior by opening /dev/fd/.  On other
systems, this may or may not work, so open_copy will fall back to
reopening the original path (and raising an exception if that path now
points to a different inode than the original file.)