WorkerProcess: do not close existing file descriptors.
The original intention here was that worker processes
(i.e. finalization) shouldn't depend on any file descriptors that were
previously opened, so we should forcibly close all of them that might
have been inherited when forking.

However, if there are still Python objects referring to those file
descriptors, this will cause problems.  In particular, if a file
object is no longer referenced but hasn't yet been garbage-collected,
then when the object eventually *is* garbage-collected it will try to
close the same FD a second time, causing havoc if that FD has since
been reused.

This was probably the cause of impossible-to-reproduce "Bad file
descriptor" errors, following commit
40d96ec39956db47787f2e683fc96ea0dc348502.